First pass at breaking into two cards
