+++ real-estate-listing/src/app/app.functions/serverless.json
@@ -1,0 +1,14 @@
+{
+  "runtime": "nodejs16.x",
+  "version": "1.0",
+  "appFunctions": {
+    "maps": {
+      "file": "maps.js",
+      "secrets": ["GOOGLE_MAPS"]
+    },
+    "walkscore": {
+      "file": "maps.js",
+      "secrets": ["WALKSCORE"]
+    }
+  }
+}
