+++ real-estate-listing/src/app/walkscore.json
@@ -1,10 +1,10 @@
 {
   "type": "crm-card",
   "data": {
-    "title": "Location Highlights",
+    "title": "Walkscore for this property",
     "location": "crm.record.tab",
     "fetch": {
-      "targetFunction": "listing-highlights",
+      "targetFunction": "walkscore",
       "objectTypes": [
         {
           "name": "p_listings",
